lab10 prelab finished
Not very clean, a very difficult lab that had me wishing I had planned its structure better. However, it works, and it does so quickly through the use of hash tables to have constant time whenever possible. The queue is implememnted as a binary tree with the invariant [P<=C] & [C>=P].